+++ lib/hexapdf/font/true_type/table.rb
@@ -63,7 +63,12 @@ module HexaPDF
 
         # Calculates the checksum for the given data.
         def self.calculate_checksum(data)
-          data.unpack('N*').inject(0) {|sum, long| sum + long } % 2**32
+          checksum = 0
+          if (remainder_length = data.length % 4) != 0
+            checksum = (data[-remainder_length, remainder_length] << "\0" * (4 - remainder_length)).
+              unpack1('N')
+          end
+          checksum + data.unpack('N*').inject(0) {|sum, long| sum + long } % 2**32
         end
 
         # The TrueType font object associated with this table.